% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{calc.R0}
\alias{calc.R0}
\title{Calculate the basic reproductive number (R0)}
\usage{
calc.R0(RM_out, population, epoch)
}
\arguments{
\item{RM_out}{named list containing the results of a Ross-Macdonald simulation (output of \code{run.RM()})}

\item{population}{either "H" or "V" for the host or vector population, respectively}

\item{epoch}{a vector of length 2 specifying the time range that should be considered - defaults to the entire simulated time period}
}
\value{
The R0 value
}
\description{
Calculates the basic reproductive number (R0) for a given Ross-Macdonald simulation for either hosts or vectors.
}
\details{
This function returns the average number of hosts that an infected host goes on to infect OR the average number of vectors that an infected vector goes on to infect. If specified, the function will only calculate this value for infections that started within a particular time range (epoch).
}
\examples{
## calculate R0 value for hosts in the example dataset
 calc.R0(RM_out = RMsim::sim3, population="H")

 # returns:

 [1] 1.000079
 calc.R0(RM_out = RMsim::sim3, population="H", epoch=c(1, 300))

 # returns:

 [1] 2.201474
}
