% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{generate.SLiM.input}
\alias{generate.SLiM.input}
\title{Generate SLiM input}
\usage{
generate.SLiM.input(infection_record)
}
\arguments{
\item{infection_record}{Infection record contained within object resulting from a Ross-Macdonald simulation after samples have been defined (output of \code{run.RM()} after being processed by \code{sample.RM()})}
}
\description{
Produces the input needed for the SLiM simulation phase from the Ross-Macdonald format generated by \code{run.RM()}.
}
\details{
This step starts with the samples resulting from the forward-in-time Ross-Macdonald simulation process (i.e. the results of \code{run.RM()}) and traces their history back to a seed infection. All infections that did not contribute to the history of the defined sample set can then be eliminated from the record. This allows later forward-in-time steps in the simulation pipeline to be more efficient, as the pruned infection record still contains all the information needed to construct the history of the sampled infections.
}
\examples{
## TBD
}
