% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_inf_histogram}
\alias{plot_inf_histogram}
\title{Plot infected proportion at equilibrium}
\usage{
plot_inf_histogram(RM_out, days_of_burn_in, scale_x_axis = TRUE, plot_hosts = TRUE, plot_vectors = TRUE, phases, observed_col = "red")
}
\arguments{
\item{RM_out}{object containing the results of a Ross-Macdonald simulation (output of \code{run.RM()})}

\item{days_of_burn_in}{the number of time steps to discard from the start of the simulation phase in calculating the empirical mean and standard deviation}

\item{scale_x_axis}{if TRUE, the x-axis is scaled according to the values in RM_out. If FALSE, the x-axis range is set to c(0,1). Default is TRUE.}

\item{plot_hosts}{if TRUE, values for the host population will be plotted}

\item{plot_vectors}{if TRUE, values for the vector population will be plotted}

\item{phases}{which phases of the Ross-Macdonald simulation, as defined by a unique set of epidemiological parameters, should be plotted. Each phase is plotted separately.}

\item{phases_per_plot}{the number of phases to plot on the same plotting space, allowing multiple phases to be visualized side by side. Default is 1.}

\item{observed_col}{the colour to plot values related to the empirical distribution}
}
\description{
Generates a histogram of the actual proportions of the host and vector populations that are infected following a specified burn-in period.
}
\details{
Parses the result of \code{run.RM()} to plot a histogram of the proportion of host and vector population that are infected at each time step of the simulation. Note that these data will be substantially auto-correlated. Nevertheless, this function allows users to visualize how well the results of the stochastic simulation match with the theoretical expectation.\cr\cr
The \code{days_of_burn_in} and \code{phases} arguments allow users to specify the time period to be plotted. \code{days_of_burn_in} should be chosen so as to exclude the initial phase of a simulation when equilibrium has not yet been reached. Viewing the result of \code{\link[=plot_inf_over_time]{plot_inf_over_time()}} can help users choose this value.\cr\cr
Values related to the empirical distribution are plotted in the colour specified by \code{observed_col}; the dashed vertical line represents the mean, and the solid lines represent 1 standard deviation above and below the mean. The expected infected proportion of the population, calculated from the underlying Ross-Macdonald equations, is plotted in black.
}
\examples{
## plot the results of the basic Ross-Macdonald simulation example
 plot_inf_histogram(RMsim::sim3, days_of_burn_in = 500, phases_per_plot=3)
 plot_inf_histogram(RMsim::sim3, days_of_burn_in = 500, phases_per_plot=3, scale_x_axis = FALSE)
 plot_inf_histogram(RMsim::sim3, days_of_burn_in = 500, phases_per_plot=3, plot_hosts = FALSE)
}
