% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.inf.histogram}
\alias{plot.inf.histogram}
\title{Plot proportion infected at equilibrium}
\usage{
asdf
}
\arguments{
\item{RM_out}{object containing the results of a Ross-Macdonald simulation (output of \code{run.RM()})}

\item{days_of_burn_in}{description}

\item{scale_x_axis}{description}

\item{plot_hosts}{description}

\item{plot_vectors}{description}

\item{phases}{description}

\item{observed_col}{description}
}
\description{
asdf
}
\details{
asdf
asdf.\cr\cr
asdf.
}
\examples{
## plot the results of the basic Ross-Macdonald simulation example
 plot.inf.histogram(RM_out_basic_sim, days_of_burn_in = 400)
 plot.inf.histogram(RM_out_basic_sim, scale_y_axis = F)
 plot.inf.histogram(RM_out_basic_sim, plot_hosts = F)
}
