% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_inf.over.time}
\alias{plot_inf.over.time}
\title{Plot infected proportion over time}
\usage{
plot_inf.over.time(RM_out, scale_y_axis = T, plot_hosts = T, plot_vectors = T, h_col = "#9A5EA1", v_col = "#98823C")
}
\arguments{
\item{RM_out}{object containing the results of a Ross-Macdonald simulation (output of \code{run.RM()})}

\item{scale_y_axis}{if TRUE, the y-axis is scaled according to the values in RM_out. If FALSE, the y-axis range is set to c(0,1). Default is TRUE.}

\item{plot_hosts}{if TRUE, the host values are plotted. Default is TRUE.}

\item{plot_vectors}{if TRUE, the vector values are plotted. Default is TRUE.}

\item{h_col}{desired colour for plotting the host data}

\item{v_col}{desired colour for plotting the vector data}
}
\description{
Generates a plot of the actual and expected proportions of the host and vector populations that are infected over the duration of the simulation.
}
\details{
Parses the result of \code{run.RM()} to plot the proportion of host and vector population that are infected at each time step of the simulation. The expected values of these parameters, calculated from the underlying Ross-Macdonald equations, are also plotted to facilitate comparison.
}
\examples{
## plot the results of the basic Ross-Macdonald simulation example
 plot_inf.over.time(RMsim::RM_out_basic_sim)
 plot_inf.over.time(RMsim::RM_out_basic_sim, scale_y_axis = F)
 plot_inf.over.time(RMsim::RM_out_basic_sim, plot_hosts = F)
}
